requirements/dev.txt
-r base.txt

bpython==0.16
django-debug-toolbar==1.9.1
sqlparse>=2.4
fabric3<2
pytest-django==3.1.2
coverage==4.4.1

flake8==3.5.0
flake8-per-file-ignores==0.4
